requirements.txt
Django==5.0.6
django-environ==0.11.2
django-crispy-forms==2.1
crispy-bootstrap5==2024.2
django-allauth==0.63.3
Pillow==10.3.0
psycopg2-binary==2.9.9
whitenoise==6.7.0
gunicorn==22.0.0
celery==5.4.0
redis==5.0.6
django-celery-beat==2.6.0
django-celery-results==2.5.1
reportlab==4.2.0
openpyxl==3.1.3
django-import-export==4.0.4
django-filter==24.2
django-tables2==2.7.0
django-htmx==1.18.0
django-widget-tweaks==1.5.0
phonenumbers==8.13.39
django-phonenumber-field==7.3.0
qrcode==7.4.2
python-decouple==3.8
dj-database-url==2.2.0
pandas==2.2.2
xlrd==2.0.1
